vim-patch:8.1.0953: a very long file is truncated at 2^31 lines
Problem:    A very long file is truncated at 2^31 lines.
Solution:   Use LONG_MAX for MAXLNUM. (Dominique Pelle, closes vim/vim#4011)
https://github.com/vim/vim/commit/c9629251a634d4f5988c8162ba8249026d1af687

+++ src/nvim/ex_cmds2.c
@@ -2118,7 +2118,7 @@ void ex_listdo(exarg_T *eap)
         i = (int)eap->line1 - 1;
         if (eap->addr_count <= 0) {
           // Default to all quickfix/location list entries.
-          assert(qf_size < MAXLNUM);
+          assert(qf_size < (uint64_t)MAXLNUM);
           eap->line2 = (linenr_T)qf_size;
         }
       }
